requirements.txt
discord.py
asyncio
random
json
requests
jokeapi